% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecovia_data}
\alias{ecovia_data}
\title{ECOVIA data}
\format{
A data frame with 1743 rows and 4 columns:
\describe{
  \item{id}{Subjects, in which time points are nested in}
  \item{time}{Day of intervention}
  \item{condition}{Condition (control vs. intervention)}
  \item{CO2}{Daily dietary carbon emissions}
}
}
\source{
<https://osf.io/qd7vw/?view_only=cc22e0d1de8844e1850d8ef1442fbecb>
}
\usage{
ecovia_data
}
\description{
A subset of data from the ECOVIA app
}
\keyword{datasets}
