% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{jsp_data}
\alias{jsp_data}
\title{JSP data}
\format{
A data frame with 728 rows and 7 columns:
\describe{
  \item{math_score_y1}{Score on math test year 1}
  \item{math_score_y3}{Score on math test year 3}
  \item{gender}{Gender of subjects (pupils)}
  \item{social_class}{Social class of subjects (pupils)}
  \item{school}{School, in which subjects (pupils) are nested in}
  \item{normal_score_y1}{Normal test score year 1}
  \item{normal_score_y3}{Normal test score year 3}
}
}
\source{
<https://www.bristol.ac.uk/cmm/team/hg/msm-3rd-ed/datasets.html>
}
\usage{
jsp_data
}
\description{
A subset of JSP secondary school exam data
}
\keyword{datasets}
